Componentização inicial do profile
