+++ frontend/src/routes/Timeline/ProfilePage.js
@@ -4,66 +4,31 @@ import Yeet from "../../components/Yeet.js";
 import ButtonCommon from "../../components/ButtonCommon.js";
 import axios from 'axios';
 
+import Profile from "./Profile.js";
+
 
 function ProfilePage() {
-    const [tweets, setTweets] = useState([]);
-    const [following, setFollowing] = useState(2);
-    const [followers, setFollowers] = useState(5);
+    const [currentScreen, setCurrentScreen] = useState('start');
+    const [email, setEmail] = useState('');
+    const [password, setPassword] = useState('');
 
-    useEffect(() => {
-        axios.get('http://localhost:3001/api/tweets/author/2')
-            .then(response => {
-                const sortedTweets = response.data.sort((a, b) => b.id - a.id);
-                setTweets(sortedTweets);
-            })
-            .catch(error => console.error('Error fetching tweets:', error));
-    }, []);
-
-    // Requisições dos seguidores 
-
-
-    return (
-        <div className="flex flex-col h-full w-full border-solid border-gray-600 border-b-2 border-x-[1px]">
-            <div className="flex flex-col font-thin text-xl text-white py-2">
-                <p className="text-white ms-2">User</p>
-                <p className="text-white font-bold text-sm ms-2">12 Yeets</p>
+    const renderScreen = () => {
+        if (currentScreen === 'signup') {
+            return <SignUpScreen />;
+        } else if (currentScreen === 'login') {
+            return <LoginScreen />;
+        }
+        else {
+            return <Profile />;
+        }
+    };
+        return (
+            <div className="flex h-screen">
+                {renderScreen()}
             </div>
-            <div className="flex flex-col border-y-[1px] border-gray-600 ">
-                <div className="flex relative">
-                    <img src="/assets/images/banner.jpg"></img>
-                    <div className="-bottom-[100px] left-6 absolute  flex flex-col justify-center items-center space-y-2">
-                        <img className="size-32  rounded-full" src="/assets/images/user-icon.jpg"/>
-                        <p className="text-white text-lg  font-bold">
-                            @User
-                        </p>
-                    </div>
-                    <div className="absolute -bottom-14 right-4">
-                        <ButtonCommon text="Editar perfil"/>
-                    </div>
-                    
-                </div>
-
-                
-            </div>
-            <div className="mt-28 ms-6 text-white space-y-3 mb-5 ">
-                <p>Olá! Eu estou usando o Y! A nova rede social que vai revolucionar o mundo!</p>
-                <div className="flex text-white space-x-3 mt-2">
-                    <p><b>{following}</b> Seguindo</p>
-                    <p><b>{followers}</b> Seguidores</p>
-                </div>
-            </div>
-            {tweets.map(tweet => (
-                <Yeet 
-                    key={tweet.id}
-                    User="Dankei" 
-                    Date={new Date(tweet.createdAt).toLocaleString()}
-                    Content={tweet.text}
-                />
-            ))}
             
-        </div>
-        
-    );
-}
+        );
+    }
+
 
 export default ProfilePage;
